sepparetes HelpPopUpViewController into two files
